cloud-functions/short-on-truck/requirements.txt
functions-framework==3.*
google-cloud-workflows==1.*
requests==2.*
python-dotenv==1.*